// apps/backend/src/model/kanel/public/ServiceConfiguration.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ServiceInstanceId } from './ServiceInstance';

/** Represents the table public.Service_Configuration */
export default interface ServiceConfiguration {
  service_instance_id: ServiceInstanceId;

  config: unknown;

  status: string | null;
}

/** Represents the initializer for the table public.Service_Configuration */
export interface ServiceConfigurationInitializer {
  service_instance_id: ServiceInstanceId;

  config: unknown;

  /** Default value: 'active'::text */
  status?: string | null;
}

/** Represents the mutator for the table public.Service_Configuration */
export interface ServiceConfigurationMutator {
  service_instance_id?: ServiceInstanceId;

  config?: unknown;

  status?: string | null;
}